Reload haproxy if any configuration changes on HA
In some cases, for instance, when updating from a non-SSL setup in
HAProxy to an SSL setup, we don't reload haproxy's configuration.
This is problematic since we need HAProxy to serve the certificates
and the new endpoints.

This forces the reload when puppet notices changes.

Change-Id: Ie1dd809e6beef33fadad48de55e488219fb7d686
Closes-Bug: #1636921
(cherry picked from commit c01acb0c4930ecbef7de1ed876afa6120fe2ee35)