Make sure keepalived is restarted before haproxy.
When using SSL setup for undercloud, the admin and public vip required
for ssl binding by haproxy are created by keepalived.

This makes sure that keepalived is started before haproxy and thus that
the interfaces are indeed present.

This patch also ensures this is happening for overcloud ssl
configuration. The case where another load-balancing technology other
than haproxy is used is not covered.

Closes-Bug: #1638029

Change-Id: I98cb0dcd7f389a1dd38ec8324429bfef4979aa66