Call VF configuration from udev rules
When a physical function that was allocated to a guest is released back
the system, it is not automatically brought "up" and the VF
configuration is not restored. This patch creates a file containing some
udev rules to force the VF configuration.

Note: we may find that the ifup-local script is no longer required but
this will require further testing.

Change-Id: Ie6e78730aa0a748b3b5100ab7c7bc007d8ab176d
Partial-Bug: #1639901
(cherry picked from commit 9a79bdab86224e678414fd98ba2e888836083e79)