Normalize civetweb binding address if IPv6
The civetweb binding format is IP:PORT; this change ensures the IP
is enclosed in brackets if IPv6.

To do so we add the bind_ip and bind_port parameters to the
rgw service class.

The Newton version does not include the changes for the class unit
tests as these were not available yet.

Change-Id: Ib84fa3479c2598bff7e89ad60a1c7d5f2c22c18c
Co-Authored-By: Lukas Bezdicka <social@v3.sk>
Related-Bug: #1636515
(cherry picked from commit 3f16ce2c0c0f73fc863d33022033112084a9b6c5)