Create Glance's database if glance-api is enabled
Instead of checking for glance_registry_enabled, we should be checking
for glance_api_enabled. The glance-api v1 depends on the registry, which
means the database will be created but glance-api v2 doesn't which means
that not deploying the registry would result in the glance database not
being created. On the other hand, glance-registry is never deployed
without glance-api

Change-Id: Ief25dafb65f7a043fbb3d16f1d7ef834c9947a93