Include ceph::profile::client from rgw.pp
To deploy successfully the RadosGW service on a dedicated node
it is necessary to provision on the node a CephX keyring with the
needed permissions to import the RadosGW service keyring. This
change will provision any keyring passed via client_keys.

It makes possible to deploy the CephRgw service on any custom role
without including the CephClient service.

Change-Id: I5772eeb233ca241887226145a472c7a0363249cb
Closes-Bug: #1673288