Allow to configure haproxy daemon's status
Currently we hard-code the fact that haproxy starts as a daemon.
When running haproxy in a container we need this to be configurable
because the haproxy process will be pid number 1.

We are not changing the current semantics which have the 'daemon'
option always set, but we are allowing its disabling.

Change-Id: I51c482b70731f15fee4025bbce14e46a49a49938