Adds OpenDaylight resources for configuring cluster with docker
In puppet-opendaylight configuring a cluster is done using an exec
calling an opendaylight script to configure the clustering file.  For
containerizing ODL, we cannot use exec as a puppet tag, therefore we
need to use a defined type to tag with.  This patch adds that defined
type to configure a cluster and a class to call it.  These will only be
used during containerized deploys and calls the same configuration
script used by puppet-odl.

Change-Id: I020550ede0ef981582392cf6c48dd5cb5823a074
Signed-off-by: Tim Rozet <trozet@redhat.com>