Change CRL refresh to run every 2 hours
The default CA issues CRLs for 4 hours by default. So we need to change
these values to reflect this, else we'll get verification issues due to
the CRL having expired before its refreshed.

However, the nextupdate value for the CRLs might not be aligned with the
cron job. And getting this alignment is not entirely trivial. So I opted
for updating every 2 hours to address this.

Change-Id: I732b400462c5cabd7c6c18c007fc9e8c87b700d3