Always start httpd at the same time
Puppet wipes out whatever is not in it's resource catalog each run for
httpd. This causes httpd to restart if in the next step there are
reasources added that were not there earlier.

This patch, thus changes the instances of httpd to start at the same
time: On step 3 for the bootstrap node, and on step 4 for every other
node.

 Conflicts:
	manifests/profile/base/heat/api.pp
	manifests/profile/base/heat/api_cfn.pp
	manifests/profile/base/heat/api_cloudwatch.pp
	manifests/profile/base/horizon.pp
	manifests/profile/base/zaqar.pp

Closes-Bug: #1699502
Change-Id: I3d29728c1ab7bd5b78100f89e00e5fa082f97b0c
Co-Authored-By: Alex Schultz <aschultz@redhat.com>
(cherry picked from commit ad14f23c112922399b96601806da34ad045cbf22)